Fixed some component work and setup form
Need to fix tournament config and get details. Need to redo the logic in the backend where the matches and the tournament are being initialized. Then, need to get the details figured out and store them. Also, need to add date property to the match and tournament as well as the setup